Clean up a bit and permit to return to hub
