+++ msvc/test/test.vcxproj
@@ -173,6 +173,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\example\DemoLevel.hpp" />
+    <ClInclude Include="..\..\example\DemoUtils.hpp" />
     <ClInclude Include="..\..\example\TestLevel.hpp" />
     <ClInclude Include="stdafx.h" />
   </ItemGroup>
